refactor!(conftest): Move to tmuxp/conftest.py for doctest
This allows pytest to detect the conftest module
